Lulz

+++ worker/mem.cpp
@@ -1,9 +1,8 @@
+#include <Windows.h>
 #include "mem.h"
-#include <windows.h>
-#include <string.h>
 
-void mem_write(u8 *dst, u8 *src, u32 len) {
-	u32 old_protect, new_protect;
+void mem_write(u8 *dst, u8 *src, size_t len) {
+	DWORD old_protect, new_protect;
 
 	VirtualProtect(dst, len, PAGE_EXECUTE_READWRITE, &old_protect);
 	while (len--) *dst = *src;
